+++ dart/packages/fory/lib/src/codegen/analyze/impl/annotation/enum_annotation_analyzer.dart
@@ -25,7 +25,7 @@ import 'package:fory/src/codegen/analyze/annotation/require_location_level.dart'
 import 'package:fory/src/codegen/const/location_level.dart';
 import 'package:fory/src/codegen/entity/location_mark.dart';
 import 'package:fory/src/annotation/fory_enum.dart';
-import 'package:fory/src/codegen/exception/annotation_exception.dart';
+import 'package:fory/src/fory_exception.dart';
 
 class EnumAnnotationAnalyzer {
   const EnumAnnotationAnalyzer();
